mcroblog_back_end/requirements.txt
django==1.10.3
djangorestframework==3.5.3
Pillow==3.4.2
python-dateutil==2.6.0
redis==2.10.5
django-cors-headers==1.1.0